Theoretically works. Needs to be tested. Code is uploaded now into the pico on orca
